// src/app/interfaces/sample-request.ts
export interface SampleRequest {
  // sampleNumber and reportNumber will be auto-generated by backend
  projectName: string;
  address: string;
  sampleDescription: string;
  samplingAndAnalysisProtocol: string;
  formatNumber: string;
  partyReferenceNumber: string;
  reportingDate: string;
  periodOfAnalysis: string;
  dateOfReceiving: string;
}

export interface RestApiResponse {
  success: boolean;
  message: string;
  data?: any;
  timestamp?: string;
}

export interface GeneralInformation {
  id?: number;
  name: string;
  value: string;
  reportNumber: string;
}